requirements.txt
langchain-google-genai==2.1.2
langgraph == 0.3.21
langgraph-prebuilt==0.1.7
langchain-community
pydantic
matplotlib
seaborn
plotly
nbformat
ipython
